# perf: speedup id construction time