--- a/docarray/document/data.py
+++ b/docarray/document/data.py
@@ -35,7 +35,9 @@ _all_mime_types = set(mimetypes.types_map.values())
 @dataclass(unsafe_hash=True)
 class DocumentData:
     _reference_doc: 'Document' = field(hash=False, compare=False)
-    id: str = field(default_factory=random.getrandbits(128).to_bytes(16, 'big').hex)
+    id: str = field(
+        default_factory=lambda: random.getrandbits(128).to_bytes(16, 'big').hex()
+    )
     parent_id: Optional[str] = None
     granularity: Optional[int] = None
     adjacency: Optional[int] = None