In this PR
```python
In [4]: %timeit da = DocumentArray.empty(100_000)
396 ms ± 3.59 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

In main ( with the last version of `os.urandom(16).hex()`)
```python
In [3]: %timeit da = DocumentArray.empty(100_000)
603 ms ± 7.34 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Note that I don't simply do `hex(random.getrandbits(128))` because this does not warranty that the hex value will be 32 hex long, which would not make it parsable as a UUID. 

This has a very high chance of failing
```python
uuids = [uuid.UUID(hex(random.getrandbits(128))) for i in range(100_000)]
```
whereas this will allways work
```python
uuids = [uuid.UUID(random.getrandbits(128).to_bytes(16, 'big').hex()) for i in range(100_000)]
```

Note: in future versions, when we use Python 3.9 , we can use  `random.randbytes` which would allow us to generate this simply with `random.randbytes(16)`.