@hanxiao 2022-03-10
> Note that I don't simply do hex(random.getrandbits(128)) because this does not warranty that the hex value will be 32 hex long, which would not make it parsable as a UUID.

Since when we promise that?