@davidbp 2022-03-10
We 

> > Note that I don't simply do hex(random.getrandbits(128)) because this does not warranty that the hex value will be 32 hex long, which would not make it parsable as a UUID.
> 
> Since when we promise that?

Never, but I find it inconvenient if users  want to parse it as UUID or we need to do it to specify a UUID for a storage backend. There is a vewy low diference in speed